Add navigation to user profile via profile image click

+++ src/main/java/ru/kpfu/itis/Kolodeznikova/servlet/requests/respondents/ConfirmRespondentServlet.java
@@ -53,7 +53,9 @@ public class ConfirmRespondentServlet extends HttpServlet {
             workoutService.createWorkout(workout, request);
 
             for (Integer id : request.getRespondentsId()) {
-                workoutRequestService.deleteRespondentFromRequest(request, id, false);
+                if (id != workout.getId()) {
+                    workoutRequestService.deleteRespondentFromRequest(request, id, false);
+                }
             }
             workoutRequestService.deleteWorkoutRequest(request, false);
